Remove bootstrap-python role from run-integration-tests pre phase
In the end, we fail regardless, ansible -m setup is run before
the job and as the cloud images does not have Python we don't reach
to run-integration-tests execution at all.
We need to either pre-bake Python on the Ubuntu nodes for bastions
or re-architect the jobs to not use bastion, I believe that will require
moving run-integration-tests to the config-project to make it trusted.